Add support for multiple delegates in textview
