+++ Sources/Views/ALKChatBarTextView.swift
@@ -27,16 +27,74 @@ open class ALKChatBarTextView: UITextView {
         return super.canPerformAction(action, withSender: sender)
     }
 
+    open override var text: String! {
+        get { return super.text }
+        set {
+            let didChange = super.text != newValue
+            super.text = newValue
+            if didChange {
+                delegate?.textViewDidChange?(self)
+            }
+        }
+    }
+
     open override var attributedText: NSAttributedString! {
         get { return super.attributedText }
         set {
             let didChange = super.attributedText != newValue
             super.attributedText = newValue
             if didChange {
-                // TODO: in case of first empty it resets to min. height
-                // nothing to change here, just adding a comment
                 delegate?.textViewDidChange?(self)
             }
         }
+    }
+
+    open override var delegate: UITextViewDelegate? {
+        get { return self }
+        set {}
+    }
+
+    private let delegates: NSHashTable<UITextViewDelegate> = NSHashTable.weakObjects()
+
+    func add(delegate: UITextViewDelegate) {
+        delegates.add(delegate)
+    }
+
+    func remove(delegate: UITextViewDelegate) {
+        for oneDelegate in delegates.allObjects.reversed() {
+            if oneDelegate === delegate {
+                delegates.remove(oneDelegate)
+            }
+        }
+    }
+
+    fileprivate func invoke(invocation: (UITextViewDelegate) -> ()) {
+        for delegate in delegates.allObjects.reversed() {
+            invocation(delegate)
+        }
+    }
+}
+
+extension ALKChatBarTextView: UITextViewDelegate {
+
+    public func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange, replacementText text: String) -> Bool {
+        invoke{ _ = $0.textView?(textView, shouldChangeTextIn: range, replacementText: text) }
+        return true
+    }
+
+    public func textViewDidChangeSelection(_ textView: UITextView) {
+        invoke { $0.textViewDidChangeSelection?(textView) }
+    }
+
+    public func textViewDidBeginEditing(_ textView: UITextView) {
+        invoke { $0.textViewDidBeginEditing?(textView) }
+    }
+
+    public func textViewDidEndEditing(_ textView: UITextView) {
+        invoke { $0.textViewDidEndEditing?(textView) }
+    }
+
+    public func textViewDidChange(_ textView: UITextView) {
+        invoke { $0.textViewDidChange?(textView) }
     }
 }
